Enhance zarr backend handling: add retry logic for server disconnects and improve materialization process

+++ src/multiview_stitcher/fusion/_core.py
@@ -1005,12 +1005,13 @@ def fuse_np(
     # Transform input views
     field_ims_t = [
         transformation.transform_sim(
-            sim.astype(np.float32),
+            # sim.astype(np.float32),
+            sim,
             np.linalg.inv(param),
             output_stack_properties=output_properties,
             order=interpolation_order,
             cval=np.nan,
-        ).data
+        ).data.astype(np.float32)
         for sim, param in zip(sims, params)
     ]
     field_ims_t = np.stack(field_ims_t)
